Rename scip io file
