+++ cpmpy/tools/io/scip_formats.py
